+++ src/index.jsx
@@ -9,10 +9,13 @@ import {
 } from '@pages/collaborate'
 import { Settings } from '@pages/config/Settings'
 import { Subjkt } from '@pages/config/Subjkt'
-import { Claim } from '@pages/dao'
-import { DaoParameters } from '@pages/dao/parameters'
-import { DaoProposals } from '@pages/dao/proposals'
-import { CreateDaoProposals } from '@pages/dao/create'
+import { Claim } from '@pages/dao/claim'
+import { DAO } from '@pages/dao'
+import {
+  DaoParameters,
+  DaoProposals,
+  SubmitDaoProposals,
+} from '@pages/dao/tabs'
 import { FAQ } from '@pages/faq'
 import { Home } from '@pages/home'
 import FriendsFeed from '@pages/home/feeds/friends-feed'
@@ -140,9 +143,12 @@ const router = createBrowserRouter(
       <Route path="subjkt/*" element={<Subjkt />} />
       <Route path="settings/*" element={<Settings />} />
       <Route path="claim/*" element={<Claim />} />
-      <Route path="dao/*" element={<DaoParameters />} />
-      <Route path="proposals/*" element={<DaoProposals />} />
-      <Route path="create/*" element={<CreateDaoProposals />} />
+      <Route path="dao/*" element={<DAO />}>
+        <Route index element={<DaoParameters />} />
+        <Route path="proposals" element={<DaoProposals />} />
+        <Route path="submit" element={<SubmitDaoProposals />} />
+        <Route path="*" element={<DaoParameters />} />
+      </Route>
       <Route path="tags/:tag" element={<Tags />} />
       <Route path="tz/:address/*" element={<Display />}>
         {display_routes}
