+++ src/pages/dao/tabs/Submit.jsx
@@ -2,15 +2,14 @@ import { useState } from 'react'
 import { DAO_GOVERNANCE_CONTRACT, DAO_TOKEN_DECIMALS, TOKENS } from '@constants'
 import { useUserStore } from '@context/userStore'
 import { useDaoStore } from '@context/daoStore'
-import { Page } from '@atoms/layout'
 import { Loading } from '@atoms/loading'
 import { Button } from '@atoms/button'
 import { Line } from '@atoms/line'
 import { DaoInput, Textarea } from '@atoms/input'
 import styles from '@style'
-import { useTokenBalance, useStorage, useGovernanceParameters } from './hooks'
+import { useTokenBalance, useStorage, useGovernanceParameters } from '../hooks'
 
-export function CreateDaoProposals() {
+export function SubmitDaoProposals() {
   // Get all the required DAO information
   const daoStorage = useStorage(DAO_GOVERNANCE_CONTRACT)
   const governanceParameters = useGovernanceParameters(daoStorage)
@@ -21,11 +20,7 @@ export function CreateDaoProposals() {
 
   // Display the loading page information until all data is available
   if (!daoStorage || !governanceParameters) {
-    return (
-      <Page title="Create new DAO proposals" large>
-        <Loading message="Loading DAO information" />
-      </Page>
-    )
+    return <Loading message="Loading DAO information" />
   }
 
   // Get the minimum number of tokens needed to create proposals
@@ -35,92 +30,83 @@ export function CreateDaoProposals() {
     currentGovernanceParameters.escrow_amount / DAO_TOKEN_DECIMALS
 
   return (
-    <Page title="Create new DAO proposals" large>
-      <div className={styles.container}>
-        <div className={styles.headline}>
-          <h1>Create new DAO proposals</h1>
-        </div>
-
-        {userTokenBalance === 0 ||
-        userTokenBalance < minimumTokensToCreateProposals ? (
+    <div className={styles.container}>
+      {userTokenBalance === 0 ||
+      userTokenBalance < minimumTokensToCreateProposals ? (
+        <section className={styles.section}>
+          {userTokenBalance === 0 ? (
+            <p>Only DAO members can create proposals.</p>
+          ) : (
+            <p>
+              A minimum of {minimumTokensToCreateProposals} TEIA tokens are
+              needed to create proposals.
+            </p>
+          )}
+        </section>
+      ) : (
+        <>
           <section className={styles.section}>
-            {userTokenBalance === 0 ? (
-              <p>Only DAO members can create proposals.</p>
-            ) : (
-              <p>
-                A minimum of {minimumTokensToCreateProposals} TEIA tokens are
-                needed to create proposals.
-              </p>
-            )}
+            <h1 className={styles.section_title}>Text proposal</h1>
+            <p>
+              Use this form to create a proposal to approve a text or decission.
+            </p>
+            <p>
+              This proposal has no direct consequences on the blockchain.
+              However, if accepted and executed, it should trigger some
+              off-chain actions by one of the Teia DAO members (e.g. change a
+              website UI, decide on a dog name, buy bread at the bakery). The
+              proposal description will be stored in IPFS for archival purposes.
+            </p>
+            <TextProposalForm />
           </section>
-        ) : (
-          <>
-            <section className={styles.section}>
-              <h1 className={styles.section_title}>Text proposal</h1>
-              <p>
-                Use this form to create a proposal to approve a text or
-                decission.
-              </p>
-              <p>
-                This proposal has no direct consequences on the blockchain.
-                However, if accepted and executed, it should trigger some
-                off-chain actions by one of the Teia DAO members (e.g. change a
-                website UI, decide on a dog name, buy bread at the bakery). The
-                proposal description will be stored in IPFS for archival
-                purposes.
-              </p>
-              <TextProposalForm />
-            </section>
 
-            <section className={styles.section}>
-              <h1 className={styles.section_title}>Transfer tez proposal</h1>
-              <p>
-                Use this form to create a proposal that, if accepted, it will
-                transfer the specified amount of tez from the DAO treasury to a
-                list of tezos addresses. The proposal description will be stored
-                in IPFS for archival purposes.
-              </p>
-              <TransferTezProposalForm />
-            </section>
+          <section className={styles.section}>
+            <h1 className={styles.section_title}>Transfer tez proposal</h1>
+            <p>
+              Use this form to create a proposal that, if accepted, it will
+              transfer the specified amount of tez from the DAO treasury to a
+              list of tezos addresses. The proposal description will be stored
+              in IPFS for archival purposes.
+            </p>
+            <TransferTezProposalForm />
+          </section>
 
-            <section className={styles.section}>
-              <h1 className={styles.section_title}>Transfer token proposal</h1>
-              <p>
-                Use this form to create a proposal that, if accepted, it will
-                transfer the specified amount of token editions from the DAO
-                treasury to a list of tezos addresses. The proposal description
-                will be stored in IPFS for archival purposes.
-              </p>
-              <TransferTokenProposalForm />
-            </section>
+          <section className={styles.section}>
+            <h1 className={styles.section_title}>Transfer token proposal</h1>
+            <p>
+              Use this form to create a proposal that, if accepted, it will
+              transfer the specified amount of token editions from the DAO
+              treasury to a list of tezos addresses. The proposal description
+              will be stored in IPFS for archival purposes.
+            </p>
+            <TransferTokenProposalForm />
+          </section>
 
-            <section className={styles.section}>
-              <h1 className={styles.section_title}>Lambda function proposal</h1>
-              <p>
-                Use this form to create a proposal that, if accepted, it will
-                execute some smart contract code stored in a Michelson lambda
-                function. The proposal description will be stored in IPFS for
-                archival purposes.
-              </p>
-              <p>
-                This proposal could be used to administer other smart contracts
-                of which the DAO is the administrator (e.g. to update the Teia
-                marketplace fees), or to execute entry points from other
-                contracts (e.g. swap or collect a token, vote in anoter DAO /
-                multisig).
-              </p>
-              <p>
-                Warning: Executing arbitrary smart contract code could
-                compromise the DAO or have unexpected consequences. The lambda
-                function code should have been revised by some trusted smart
-                contract expert before the proposal is accepted and executed.
-              </p>
-              <LambdaFunctionProposalForm />
-            </section>
-          </>
-        )}
-      </div>
-    </Page>
+          <section className={styles.section}>
+            <h1 className={styles.section_title}>Lambda function proposal</h1>
+            <p>
+              Use this form to create a proposal that, if accepted, it will
+              execute some smart contract code stored in a Michelson lambda
+              function. The proposal description will be stored in IPFS for
+              archival purposes.
+            </p>
+            <p>
+              This proposal could be used to administer other smart contracts of
+              which the DAO is the administrator (e.g. to update the Teia
+              marketplace fees), or to execute entry points from other contracts
+              (e.g. swap or collect a token, vote in anoter DAO / multisig).
+            </p>
+            <p>
+              Warning: Executing arbitrary smart contract code could compromise
+              the DAO or have unexpected consequences. The lambda function code
+              should have been revised by some trusted smart contract expert
+              before the proposal is accepted and executed.
+            </p>
+            <LambdaFunctionProposalForm />
+          </section>
+        </>
+      )}
+    </div>
   )
 }
