+++ src/pages/dao/tabs/index.js
@@ -1,0 +1,3 @@
+export { DaoParameters } from './Parameters'
+export { DaoProposals } from './Proposals'
+export { SubmitDaoProposals } from './Submit'
