+++ src/pages/dao/tabs/index.module.scss
@@ -1,0 +1,254 @@
+@import '@styles/mixins.scss';
+@import '@styles/variables.scss';
+
+.container {
+  padding-top: 12px;
+  padding-bottom: 60px;
+  width: 100%;
+
+  @include respond-to('desktop') {
+    margin: 0;
+  }
+}
+
+.section_title {
+  margin-bottom: 1em;
+}
+
+.section {
+  margin: 1em 0;
+
+  b {
+    font-weight: bold;
+  }
+
+  i {
+    font-style: italic;
+  }
+
+  a {
+    text-decoration: underline;
+  }
+
+  > p {
+    margin: 1em 0;
+  }
+
+  li {
+    margin-left: 1.5em;
+    margin-top: 0.8em;
+  }
+
+  li:before {
+    content: attr(counter) '.';
+    position: absolute;
+    margin-left: -1.5em;
+  }
+
+  @include respond-to('desktop') {
+    margin-bottom: 30px;
+  }
+}
+
+.parameters_list {
+  li:before {
+    content: '';
+  }
+}
+
+.ipfs_link {
+  word-break: break-word;
+}
+
+.proposal_list {
+  max-height: 50em;
+  overflow: auto;
+
+  & > li {
+    padding: 0.2em 0;
+
+    &:before {
+      content: '';
+    }
+
+    &:not(:last-child) {
+      border-bottom: 1px solid darken(white, 30%);
+    }
+  }
+}
+
+.user_votes {
+  display: inline-block;
+  min-width: 3em;
+  padding: 0.5em;
+  border-radius: 4px;
+  background-color: white;
+  margin: 0 0.5em 0.5em 0;
+
+  &:empty:before {
+    content: '\200b';
+  }
+}
+
+.yes_vote {
+  background-color: lighten(green, 55%);
+}
+
+.no_vote {
+  background-color: lighten(red, 25%);
+}
+
+.abstain_vote {
+  background-color: lighten(yellow, 25%);
+}
+
+.proposal_information {
+  p {
+    margin: 0.5em 0;
+  }
+}
+
+.proposal_id {
+  display: inline-block;
+  background-color: white;
+  color: lighten(black, 20%);
+  min-width: 4em;
+  padding: 0 0.5em;
+  margin-right: 0.5em;
+  border-radius: 4px;
+  text-align: center;
+}
+
+details {
+  summary {
+    cursor: pointer;
+  }
+
+  table {
+    display: inline-block;
+    margin: 0.5em 0 0 0;
+    padding: 1em;
+    border-radius: 4px;
+    overflow: hidden;
+  }
+
+  td:first-child {
+    text-align: right;
+  }
+}
+
+.proposal_votes_summary {
+  margin-top: 1em;
+}
+
+.votes_display_result {
+  font-size: 0.9em;
+  text-align: center;
+  color: black;
+  margin: 0 1em;
+  display: flex;
+  flex-direction: row;
+}
+
+.vote_display_nothing {
+  background-color: white;
+}
+
+.vote_display_yes {
+  background-color: lighten(green, 55%);
+}
+
+.vote_display_no {
+  background-color: lighten(red, 25%);
+}
+
+.vote_display_abstain {
+  background-color: lighten(yellow, 25%);
+}
+
+.proposal_actions {
+  margin-right: 0.5em;
+
+  > :not(:first-child) {
+    margin-top: 1em;
+  }
+
+  p {
+    margin: 0.5em 0;
+  }
+}
+
+.micheline_code {
+  line-height: 1.4;
+  margin: 1em 0;
+  padding: 1em;
+  border-radius: 4px;
+  white-space: pre-wrap;
+  word-break: break-word;
+}
+
+.proposal {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  align-items: baseline;
+  gap: 0.5em;
+
+  @include respond-to('desktop') {
+    flex-direction: row;
+  }
+}
+
+.proposal_information {
+  width: 100%;
+  @include respond-to('desktop') {
+    flex: 1 1 0;
+  }
+}
+
+.proposal_actions_buttons {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  gap: 0.5em;
+  align-items: baseline;
+
+  @include respond-to('desktop') {
+    justify-content: flex-end;
+  }
+}
+
+.proposal_form {
+  margin: 2em 0 3em 0;
+
+  input[type='file'] {
+    display: none;
+  }
+}
+
+.proposal_form_field {
+  margin: 1em 0;
+}
+
+.proposal_description_label {
+  display: inline-block;
+  font-weight: 700;
+  letter-spacing: $letter-spacing;
+  color: inherit;
+  pointer-events: none;
+  border: 1px;
+}
+
+.upload_button {
+  display: block;
+  width: 100%;
+  margin-top: 1em;
+  padding: 6px;
+  border: 2px dashed var(--text-color);
+  text-align: center;
+  cursor: pointer;
+}
+
+.transfers_fields {
+  margin-bottom: 1em;
+}
